Finished migrations

+++ database/migrations/2019_06_29_032854_create_images_table.php
@@ -15,7 +15,7 @@ class CreateImagesTable extends Migration
     {
         Schema::create('images', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('category_id')->unsigned();
+            $table->bigInteger('category_id')->unsigned();
             $table->integer('sort_order')->unsigned()->default(0);
             $table->string('path');
             $table->string('url')->nullable();
